feat: app chrome & library sidebar layout shuffling
