+++ packages/raga-app/src/client/components/chrome/appChrome.tsx
@@ -1,16 +1,13 @@
-import { Divider, Group, Title } from "@mantine/core";
+import { Divider, Group } from "@mantine/core";
 import classNames from "classnames";
 
 import { appStore } from "../../store/appStore";
 import UserSettingsDropdown from "../settings/userSettingsDropdown";
 import styles from "./appChrome.module.scss";
-import AudioAnalyzerStatus from "./audioAnalyzerStatus";
-import AudioFilesServerControls from "./audioFilesServerControls";
 import LibraryControls from "./libraryControls";
-import WorkspaceControl from "./workspaceControl";
+import WorkspaceModeControl from "./workspaceModeControl";
 
 export default function AppChrome() {
-  const audioFilesServerStatus = appStore.use.audioFilesServerStatus();
   const isLibraryLoaded = appStore.use.libraryLoadingState() === "loaded";
   const isMacOS = window.api.platform === "darwin";
   const isWindows = window.api.platform === "win32";
@@ -35,27 +32,16 @@ export default function AppChrome() {
         gap="sm"
         className={styles.appChromeGroup}
       >
-        <Title order={4} className={styles.appChromeDraggable}>
-          Raga
-        </Title>
         {isLibraryLoaded && (
           <>
-            <Divider orientation="vertical" />
-            <WorkspaceControl />
-            <AudioFilesServerControls />
+            <WorkspaceModeControl />
           </>
         )}
-        {audioFilesServerStatus === "started" && (
-          <>
-            <Divider orientation="vertical" />
-            <AudioAnalyzerStatus />
-          </>
-        )}
+      </Group>
+      <Group className={styles.appChromeGroup} p={5} gap="xs">
+        <LibraryControls />
         <Divider orientation="vertical" />
         <UserSettingsDropdown />
-      </Group>
-      <Group className={styles.appChromeGroup} p={5}>
-        <LibraryControls />
       </Group>
     </Group>
   );
