--- packages/raga-app/src/client/components/chrome/libraryControls.module.scss
@@ -1,9 +1,0 @@
-.buttonNoRightRadius {
-  border-top-right-radius: 0;
-  border-bottom-right-radius: 0;
-}
-
-.buttonNoLeftRadius {
-  border-top-left-radius: 0;
-  border-bottom-left-radius: 0;
-}
