+++ packages/raga-app/src/client/components/playlistTable/playlistTable.module.scss
@@ -2,7 +2,6 @@
 
 .playlistTableContainer {
   padding-left: 1px; // allows border from container element to show
-  padding-bottom: 31px; // space for footer + border
   height: 100%;
   display: flex;
   flex-direction: column;
