+++ packages/raga-app/src/client/views/libraryView.module.scss
@@ -4,7 +4,6 @@
 
   .librarySidebarFooter {
     bottom: 0;
-    height: 30px;
     left: 0;
     position: absolute;
     right: 0;
